refactor(api-file-manager-server): wrap routes in feature folders
Routes moved into feature folders with feature.ts files.
Registered via feature pattern instead of bare container.register().
Removed comment explaining what code does. Separated re-exports
from registration logic.
