+++ packages/api-file-manager-server/src/routes/UploadPartRoute/feature.ts
@@ -1,0 +1,9 @@
+import { createFeature } from "@webiny/feature/api";
+import { UploadPartRoute } from "./UploadPartRoute.js";
+
+export const UploadPartRouteFeature = createFeature({
+    name: "FileManagerServer/UploadPartRoute",
+    register(container) {
+        container.register(UploadPartRoute);
+    }
+});
