+++ packages/api-file-manager-server/src/routes/UploadSingleFileRoute/feature.ts
@@ -1,0 +1,9 @@
+import { createFeature } from "@webiny/feature/api";
+import { UploadSingleFileRoute } from "./UploadSingleFileRoute.js";
+
+export const UploadSingleFileRouteFeature = createFeature({
+    name: "FileManagerServer/UploadSingleFileRoute",
+    register(container) {
+        container.register(UploadSingleFileRoute);
+    }
+});
